Added art previews loading, added art creation method

+++ app/src/main/java/ru/nsu/fit/android/drawalk/model/firebase/PointSettings.kt
@@ -2,5 +2,5 @@ package ru.nsu.fit.android.drawalk.model.firebase
 
 data class PointSettings(
     var color: Int = 0,
-    var width: Int = 1
+    var width: Double = 1.0
 )
